app/requirements.txt
langchain
langchain-community
langchain-groq
unstructured
chromadb
streamlit
pandas
python-dotenv
PyPDF2
transformers
sentence-transformers
matplotlib